stream-processing/requirements.txt
apache-flink==1.19.0
psycopg2-binary>=2.9
packaging>=24.0
